+++ src/test/java/app/getxray/xray/it/import_results/XrayCloudIT.java
@@ -11,6 +11,7 @@ import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 import com.github.tomakehurst.wiremock.junit5.WireMockTest;
 import com.soebes.itf.jupiter.extension.MavenGoal;
 import com.soebes.itf.jupiter.extension.MavenJupiterExtension;
+import com.soebes.itf.jupiter.extension.MavenOption;
 import com.soebes.itf.jupiter.extension.MavenTest;
 import com.soebes.itf.jupiter.extension.SystemProperty;
 import com.soebes.itf.jupiter.maven.MavenExecutionResult;
@@ -131,6 +132,45 @@ public class XrayCloudIT {
     @SystemProperty(value = "xray.reportFormat", content = "xray")
     @SystemProperty(value = "xray.reportFile", content = "xray.json")
     @SystemProperty(value = "xray.useInternalTestProxy", content = "true")
+    @SystemProperty(value = "xray.verbose", content = "true")
+    @MavenOption("--debug")
+    void xray_standard_with_verbose_mode(MavenExecutionResult result) throws IOException {
+        /*
+        String report = CommonUtils.readResourceFileForImportResults("XrayCloudIT/xray_standard/xray.json");
+
+        wm.verify(
+            postRequestedFor(urlPathEqualTo("/api/v2/import/execution"))
+                .withHeader("Content-Type", containing("application/json"))
+                .withRequestBody(equalToJson(report))
+        );
+       */
+        assertThat(result).isSuccessful();
+        assertThat(result)
+            .out()
+            .debug()
+            .containsOnlyOnce("REQUEST_URL: https://xray.cloud.getxray.app/api/v2/authenticate");
+        assertThat(result)
+            .out()
+            .debug()
+            .containsOnlyOnce("REQUEST_URL: https://xray.cloud.getxray.app/api/v2/import/execution");
+        assertThat(result)
+            .out()
+            .debug()
+            .contains("REQUEST_METHOD: POST");
+        assertThat(result)
+            .out()
+            .debug()
+            .contains("REQUEST_CONTENT_TYPE: application/json; charset=utf-8");
+    }
+
+    @MavenTest
+    @MavenGoal("xray:import-results")
+    @SystemProperty(value = "xray.cloud", content = "true")
+    @SystemProperty(value = "xray.clientId", content = CLIENT_ID)
+    @SystemProperty(value = "xray.clientSecret", content = CLIENT_SECRET)
+    @SystemProperty(value = "xray.reportFormat", content = "xray")
+    @SystemProperty(value = "xray.reportFile", content = "xray.json")
+    @SystemProperty(value = "xray.useInternalTestProxy", content = "true")
     void xray_standard(MavenExecutionResult result) throws IOException {
        String report = CommonUtils.readResourceFileForImportResults("XrayCloudIT/xray_standard/xray.json");
 
@@ -183,6 +223,11 @@ public class XrayCloudIT {
  @SystemProperty(value = "xray.reportFormat", content = "junit")
  @SystemProperty(value = "xray.reportFile", content = "junit.xml")
  @SystemProperty(value = "xray.projectKey", content = "CALC")
+ @SystemProperty(value = "xray.testExecKey", content = "CALC-2")
+ @SystemProperty(value = "xray.testPlanKey", content = "CALC-3")
+ @SystemProperty(value = "xray.version", content = "1.0")
+ @SystemProperty(value = "xray.revision", content = "123")
+ @SystemProperty(value = "xray.testEnvironment", content = "chrome")
  @SystemProperty(value = "xray.useInternalTestProxy", content = "true")
  void junit_standard(MavenExecutionResult result) throws IOException {
     String report = CommonUtils.readResourceFileForImportResults("XrayCloudIT/junit_standard/junit.xml");
@@ -191,6 +236,11 @@ public class XrayCloudIT {
         postRequestedFor(urlPathEqualTo("/api/v2/import/execution/junit"))
             .withHeader("Content-Type", containing("application/xml"))
             .withQueryParam("projectKey", equalTo("CALC"))
+            .withQueryParam("testExecKey", equalTo("CALC-2"))
+            .withQueryParam("testPlanKey", equalTo("CALC-3"))
+            .withQueryParam("fixVersion", equalTo("1.0"))
+            .withQueryParam("revision", equalTo("123"))
+            .withQueryParam("testEnvironments", equalTo("chrome"))
             .withRequestBody(equalToXml(report))
     );
     assertThat(result).isSuccessful();