+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- added "verbose" mode, to debug API requests/responses
+
 ### Fixed
 
 ## [0.6.0] - 2022-06-08